+++ srcs/raycast/raycast_ver.c
@@ -1,6 +1,6 @@
 #include "../../includes/cub.h"
 
-void	instanciate_pos_ver(t_map *map, int square_size)
+void	instanciate_pos_ver(t_map *map, double square_size)
 {
 	map->player.ray_vertical.is_wall = 0;
 	map->player.ray_vertical.pos_y = square_size * (map->player.pos_y + 1.0);
@@ -14,7 +14,7 @@ void	instanciate_pos_ver(t_map *map, int square_size)
 	}
 }
 
-void	instanciate_length_ver(t_map *map, double tang, int square_size)
+void	instanciate_length_ver(t_map *map, double tang, double square_size)
 {
 	double	floor_pos_x;
 
@@ -42,16 +42,16 @@ void	instanciate_length_ver(t_map *map, double tang, int square_size)
 }
 
 void	vertical_check(t_map *map, double *length_case_x,
-		double *length_case_y, int square)
+		double *length_case_y, double square)
 {
 	if (map->lines == NULL)
 		close_program(map, "The engine couldn't read the map.", 2);
 	check_loop_cast_ver(map);
 	if (map->player.ray_vertical.is_wall == 1
-			|| map->lines[(int)floor(*length_case_y) / square]
-			[(int)floor(*length_case_x) / square] == '1'
-			|| map->lines[(int)floor(*length_case_y) / square]
-			[(int)floor(*length_case_x) / square] == ' ')
+			|| map->lines[(int)floor(*length_case_y) / (int)square]
+			[(int)floor(*length_case_x) / (int)square] == '1'
+			|| map->lines[(int)floor(*length_case_y) / (int)square]
+			[(int)floor(*length_case_x) / (int)square] == ' ')
 	{
 		map->player.ray_vertical.distance_wall =
 			sqrt(pow(map->player.ray_vertical.pos_x
@@ -71,9 +71,9 @@ void	vertical_check(t_map *map, double *length_case_x,
 	}
 }
 
-void	vertical_detection(t_map *map, int number_lines, int square_size)
+void	vertical_detection(t_map *map, int number_lines, double square_size)
 {
-	int		square;
+	double	square;
 	double	tang;
 	double	length_case_y;
 	double	length_case_x;
@@ -92,9 +92,9 @@ void	vertical_detection(t_map *map, int number_lines, int square_size)
 				|| (int)floor(length_case_y / square) > number_lines
 				|| cpr_equal((int)floor(length_case_y / square), number_lines)
 				|| (length_case_x / square)
-				> max_case(map->lines[(int)floor(length_case_y / square)])
+				> max_case(map->lines[(int)floor(length_case_y / (int)square)])
 				|| cpr_equal((length_case_x / square),
-					max_case(map->lines[(int)floor(length_case_y / square)])))
+					max_case(map->lines[(int)floor(length_case_y / (int)square)])))
 			map->player.ray_vertical.is_wall = 1;
 		vertical_check(map, &length_case_x, &length_case_y, square);
 	}
